# [core] don't specify ORDER BY when counting total number of results, which allows the DB to be much faster